Move remained funasr logic to dedicated place
