+++ optimum/intel/openvino/modeling_seq2seq.py
@@ -13,7 +13,6 @@
 #  limitations under the License.
 import logging
 import os
-import shutil
 from pathlib import Path
 from typing import Dict, List, Optional, Tuple, Union
 
@@ -1323,247 +1322,22 @@ class OVModelForSpeechSeq2Seq(OVModelForSeq2SeqLM):
 
     @classmethod
     def from_pretrained(cls, model_id, export: bool = False, config: Optional["PretrainedConfig"] = None, **kwargs):
-        # FunASR models (e.g. Fun-ASR-Nano) are loaded via the funasr library and have no root
-        # config.json; transformers' AutoConfig / optimum's library inference cannot handle them.
-        # We detect them (either the original funasr repo, or an already-exported OpenVINO dir whose
-        # config.json carries `export_model_type == "fun_asr"`) and load the config ourselves,
-        # bypassing the standard inference path.
-        if config is None and cls._is_funasr_source(model_id, **kwargs):
-            return cls._from_pretrained_funasr(model_id, export=export, **kwargs)
+        from .modeling_funasr import _get_funasr_class, _is_funasr_source
+
+        # the original FunASR model has no config file, so _from_pretrained() dispatch does not work
+        if config is None and _is_funasr_source(model_id, **kwargs):
+            return _get_funasr_class()._from_pretrained_funasr(model_id, export=export, **kwargs)
 
         return super().from_pretrained(model_id, export=export, config=config, **kwargs)
-
-    def _save_pretrained(self, save_directory: Union[str, Path]):
-        super()._save_pretrained(save_directory)
-        # FunASR has no transformers processor: its tokenizer/detokenizer IR (and the source tokenizer
-        # files) are exported alongside the model and are required by preprocess_input. Persist them.
-        if getattr(self.config, "model_type", None) == "fun_asr" and self.model_save_dir is not None:
-            src_dir = Path(self.model_save_dir)
-            save_directory = Path(save_directory)
-            tokenizer_assets = [
-                OV_TOKENIZER_NAME.format(""),
-                OV_TOKENIZER_NAME.format("").replace(".xml", ".bin"),
-                "openvino_detokenizer.xml",
-                "openvino_detokenizer.bin",
-            ]
-            for name in tokenizer_assets:
-                src = src_dir / name
-                if src.is_file() and src.resolve() != (save_directory / name).resolve():
-                    shutil.copyfile(src, save_directory / name)
-
-    @staticmethod
-    def _is_funasr_source(model_id, **kwargs) -> bool:
-        from optimum.exporters.tasks import TasksManager
-
-        from .modeling_funasr import _is_funasr_model
-
-        cache_dir = kwargs.get("cache_dir", HUGGINGFACE_HUB_CACHE)
-        token = kwargs.get("token")
-        subfolder = kwargs.get("subfolder", "")
-        revision = kwargs.get("revision")
-        try:
-            all_files, _ = TasksManager.get_model_files(
-                model_id, subfolder=subfolder, cache_dir=cache_dir, revision=revision, token=token
-            )
-        except Exception:
-            all_files = []
-
-        # Original funasr repo (configuration.json + config.yaml, no root config.json).
-        if _is_funasr_model(model_id, all_files, cache_dir=cache_dir, token=token):
-            return True
-
-        # Previously exported OpenVINO model: config.json with export_model_type == "fun_asr".
-        if "config.json" in all_files:
-            try:
-                cfg = PretrainedConfig.from_pretrained(
-                    model_id, subfolder=subfolder, cache_dir=cache_dir, revision=revision, token=token
-                )
-                return getattr(cfg, "export_model_type", None) == "fun_asr"
-            except Exception:
-                return False
-        return False
-
-    @classmethod
-    def _from_pretrained_funasr(cls, model_id, export: bool = False, **kwargs):
-        """Loading entrypoint for FunASR models, bypassing optimum's library/config inference."""
-        # Determine whether OpenVINO IR already exists (load) or we need to export.
-        _export = export
-        try:
-            ov_files = _find_files_matching_pattern(
-                model_id,
-                pattern=cls._search_pattern,
-                subfolder=kwargs.get("subfolder", ""),
-                use_auth_token=kwargs.get("token"),
-                revision=kwargs.get("revision"),
-            )
-            _export = len(ov_files) == 0
-        except Exception:
-            pass
-
-        if _export:
-            # Build the export-time config directly from the funasr model.
-            from .modeling_funasr import _FunASRForSpeechSeq2Seq
-
-            funasr_wrapped = _FunASRForSpeechSeq2Seq.from_pretrained(
-                model_id, cache_dir=kwargs.get("cache_dir", HUGGINGFACE_HUB_CACHE), token=kwargs.get("token")
-            )
-            config = funasr_wrapped.config
-            del funasr_wrapped
-            return cls._export(model_id, config=config, **kwargs)
-
-        # Loading a previously exported model. PretrainedConfig does not serialize the instance-level
-        # `model_type` (it is a class attribute), so we restore it from `export_model_type`.
-        config = PretrainedConfig.from_pretrained(model_id)
-        if getattr(config, "export_model_type", None) == "fun_asr":
-            config.model_type = "fun_asr"
-        config.is_encoder_decoder = True
-        return cls._from_pretrained(model_id, config=config, **kwargs)
-
-    def preprocess_input(self, *args, **kwargs) -> Dict[str, torch.Tensor]:
-        """
-        Prepare raw inputs into the tensors expected by `generate()`.
-
-        This is a model-specific hook: speech models that do not ship a transformers processor
-        (e.g. FunASR) implement it to turn a waveform into `input_features` / `decoder_input_ids`.
-        The returned dictionary can be passed directly to `generate(**inputs)`.
-
-        Returns:
-            Dictionary with model inputs ready for `generate()`.
-        """
-        if getattr(self.config, "model_type", None) == "fun_asr":
-            return self._preprocess_input_funasr(*args, **kwargs)
-        raise NotImplementedError(
-            f"`preprocess_input` is not implemented for model type " f"`{getattr(self.config, 'model_type', None)}`."
-        )
-
-    def _preprocess_input_funasr(
-        self,
-        waveforms: Union[np.ndarray, torch.Tensor, List],
-        sampling_rate: int,
-        language: str = "中文",
-        itn: bool = True,
-    ) -> Dict[str, torch.Tensor]:
-        """
-        Standalone FunASR preprocessing (no `funasr` dependency).
-
-        Reproduces FunASR's WavFrontend fbank+LFR features and the chat-style prompt with audio
-        placeholder tokens, using the OpenVINO tokenizer IR exported alongside the model.
-
-        Args:
-            waveforms: A single mono waveform (1-D array/tensor) or a batch (list / 2-D array).
-            sampling_rate: Sampling rate of `waveforms` in Hz.
-            language: FunASR prompt language tag (a Chinese word, e.g. "中文", "英文", "日文").
-            itn: Whether to apply inverse text normalization.
-
-        Returns:
-            Dictionary with `input_features` (B, T, 560) and `decoder_input_ids` (B, L).
-        """
-        import torchaudio
-        import torchaudio.compliance.kaldi as kaldi
-
-        # WavFrontend configuration (from the model's config.yaml frontend_conf).
-        target_fs, n_mels, frame_length, frame_shift, lfr_m, lfr_n = 16000, 80, 25, 10, 7, 6
-        audio_token_id = getattr(self.config, "audio_token_id", 0)
-
-        def _apply_lfr(inputs: torch.Tensor) -> torch.Tensor:
-            T = inputs.shape[0]
-            T_lfr = int(np.ceil(T / lfr_n))
-            left_padding = inputs[0].repeat((lfr_m - 1) // 2, 1)
-            inputs = torch.vstack((left_padding, inputs))
-            T = T + (lfr_m - 1) // 2
-            feat_dim = inputs.shape[-1]
-            strides = (lfr_n * feat_dim, 1)
-            sizes = (T_lfr, lfr_m * feat_dim)
-            last_idx = (T - lfr_m) // lfr_n + 1
-            num_padding = lfr_m - (T - last_idx * lfr_n)
-            if num_padding > 0:
-                num_padding = (2 * lfr_m - 2 * T + (T_lfr - 1 + last_idx) * lfr_n) / 2 * (T_lfr - last_idx)
-                inputs = torch.vstack([inputs] + [inputs[-1:]] * int(num_padding))
-            return inputs.as_strided(sizes, strides).clone().type(torch.float32)
-
-        def _extract_features(waveform: torch.Tensor) -> torch.Tensor:
-            if waveform.ndim > 1:  # reduce channels to mono
-                waveform = waveform.mean(0)
-            if sampling_rate != target_fs:
-                waveform = torchaudio.transforms.Resample(sampling_rate, target_fs)(waveform[None, :])[0, :]
-            wav = waveform.float() * (1 << 15)  # upscale_samples
-            wav = wav.unsqueeze(0)
-            mat = kaldi.fbank(
-                wav,
-                num_mel_bins=n_mels,
-                frame_length=min(frame_length, wav.shape[1] / target_fs * 1000),
-                frame_shift=frame_shift,
-                dither=0.0,
-                energy_floor=0.0,
-                window_type="hamming",
-                sample_frequency=target_fs,
-                snip_edges=True,
-            )
-            return _apply_lfr(mat)  # (T, 560)
-
-        def _num_audio_tokens(num_frames: int) -> int:
-            # use_low_frame_rate=True formula from FunASRNano.data_load_speech.
-            olens = 1 + (num_frames - 3 + 2 * 1) // 2
-            olens = 1 + (olens - 3 + 2 * 1) // 2
-            return (olens - 1) // 2 + 1
-
-        # Normalize input into a list of 1-D tensors.
-        if isinstance(waveforms, (list, tuple)):
-            wavs = [torch.as_tensor(np.asarray(w)) for w in waveforms]
-        else:
-            arr = waveforms if isinstance(waveforms, torch.Tensor) else torch.as_tensor(np.asarray(waveforms))
-            wavs = [arr] if arr.ndim == 1 else list(arr)
-
-        feats = [_extract_features(w) for w in wavs]
-        num_frames = [f.shape[0] for f in feats]
-        max_frames = max(num_frames)
-        feature_size = feats[0].shape[-1]
-        input_features = torch.zeros(len(feats), max_frames, feature_size, dtype=torch.float32)
-        for i, f in enumerate(feats):
-            input_features[i, : f.shape[0]] = f
-
-        # Build the chat prompt and tokenize it with the exported OpenVINO tokenizer IR.
-        asr_prompt = f"语音转写成{language}：" if itn else f"语音转写成{language}，不进行文本规整："
-        before = f"<|im_start|>system\nYou are a helpful assistant.<|im_end|>\n<|im_start|>user\n{asr_prompt}"
-        after = "<|im_end|>\n<|im_start|>assistant\n"
-        before_ids = self._funasr_tokenizer_encode(before)
-        after_ids = self._funasr_tokenizer_encode(after)
-
-        prompt_ids = []
-        for nf in num_frames:
-            ids = before_ids + [audio_token_id] * _num_audio_tokens(nf) + after_ids
-            prompt_ids.append(torch.tensor(ids, dtype=torch.long))
-        max_len = max(t.shape[0] for t in prompt_ids)
-        decoder_input_ids = torch.zeros(len(prompt_ids), max_len, dtype=torch.long)
-        for i, t in enumerate(prompt_ids):
-            decoder_input_ids[i, : t.shape[0]] = t
-
-        return {"input_features": input_features, "decoder_input_ids": decoder_input_ids}
-
-    def _funasr_tokenizer_encode(self, text: str) -> List[int]:
-        """Encode text to token ids using the exported OpenVINO tokenizer IR."""
-        if getattr(self, "_ov_tokenizer", None) is None:
-            import openvino_tokenizers  # noqa: F401  (registers the tokenizer ops)
-
-            tokenizer_path = Path(self.model_save_dir) / OV_TOKENIZER_NAME.format("")
-            if not tokenizer_path.is_file():
-                raise FileNotFoundError(
-                    f"OpenVINO tokenizer IR not found at {tokenizer_path}. Re-export the model so the "
-                    "tokenizer/detokenizer IR is generated."
-                )
-            self._ov_tokenizer = Core().compile_model(str(tokenizer_path), "CPU")
-        result = self._ov_tokenizer([text])
-        return result["input_ids"][0].tolist()
 
     def _prepare_decoder_input_ids_for_generation(
         self, batch_size, model_input_name, model_kwargs, decoder_start_token_id, device=None
     ):
         """
-        For qwen3_asr / fun_asr: skip prepending decoder_start_token_id since the full prompt
+        For qwen3_asr: skip prepending decoder_start_token_id since the full prompt
         (including chat template tokens) is already provided as decoder_input_ids.
-        This matches the PyTorch model behavior where input_ids is used as-is.
         """
-        if getattr(self.config, "model_type", None) in ("qwen3_asr", "fun_asr"):
+        if getattr(self.config, "model_type", None) == "qwen3_asr":
             if model_kwargs is not None and "decoder_input_ids" in model_kwargs:
                 decoder_input_ids = model_kwargs.pop("decoder_input_ids")
             elif "input_ids" in model_kwargs and model_input_name != "input_ids":
@@ -1572,11 +1346,9 @@ class OVModelForSpeechSeq2Seq(OVModelForSeq2SeqLM):
                 decoder_input_ids = None
 
             if decoder_input_ids is None:
-                # Fallback to default behavior if no decoder_input_ids provided
                 return super()._prepare_decoder_input_ids_for_generation(
                     batch_size, model_input_name, model_kwargs, decoder_start_token_id, device
                 )
-            # Return decoder_input_ids as-is without prepending decoder_start_token_id
             return decoder_input_ids, model_kwargs
 
         return super()._prepare_decoder_input_ids_for_generation(
@@ -1720,12 +1492,14 @@ class OVModelForSpeechSeq2Seq(OVModelForSeq2SeqLM):
     ):
         if "WhisperForConditionalGeneration" in (getattr(config, "architectures", None) or []):
             return _OVModelForWhisper._from_pretrained(model_id, config, **kwargs)
-        elif getattr(config, "model_type", None) in ("qwen3_asr", "fun_asr"):
-            # Ensure is_encoder_decoder is set for proper model loading
+        if getattr(config, "model_type", None) == "fun_asr":
+            from .modeling_funasr import _get_funasr_class
+
             config.is_encoder_decoder = True
-            return super()._from_pretrained(model_id, config, **kwargs)
-        else:
-            return super()._from_pretrained(model_id, config, **kwargs)
+            return _get_funasr_class()._from_pretrained(model_id, config, **kwargs)
+        if getattr(config, "model_type", None) == "qwen3_asr":
+            config.is_encoder_decoder = True
+        return super()._from_pretrained(model_id, config, **kwargs)
 
 
 class _OVModelForWhisper(OVModelForSpeechSeq2Seq, WhisperForConditionalGeneration):
